+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2020, Swann Martinez'
 author = 'Swann Martinez'
 
 # The full version, including alpha/beta/rc tags
-release = '0.0.2'
+release = '0.1.0'
 
 
 # -- General configuration ---------------------------------------------------