+++ tests/test_bl_types/test_scene.py
@@ -6,8 +6,11 @@ from deepdiff import DeepDiff
 import bpy
 import random
 from multi_user.bl_types.bl_scene import BlScene
+from multi_user.utils import get_preferences
 
 def test_scene(clear_blend):
+    get_preferences().sync_flags.sync_render_settings = True
+
     datablock = bpy.data.scenes.new("toto")
     datablock.view_settings.use_curve_mapping = True
     # Test